interop: move the last unpinned FFI enums where a test can see them
Eleven C enums lived beside the P/Invokes in the WinUI project: platform,
surface context, clipboard and clipboard request, mouse state and button,
colour scheme, mods, input action, point tag and point coord. Ghostty.Tests
cannot reference that project, so nothing checked any of them, and they are
the same shape as the ones that just broke: positions in an enum edited
elsewhere, crossing as ints, renumbered by an insertion without breaking a
build. Mouse button has twelve members and point tag four.

Moved to Ghostty.Core beside the action payload enums, which is where the
existing pinned ones already live, and pinned against include/ghostty.h in
both directions.

Three things the header made the checker learn on the way. Values written
`1 << N` now parse, so mods reads as bit flags rather than being refused.
The [Flags] zero member is only exempt when the header does not name one:
ghostty_input_mods_e does spell GHOSTTY_MODS_NONE, and a blanket exemption
would have stopped the completeness check noticing if it went away. And the
name mapping breaks at a letter-to-digit boundary as well as at a capital,
because the header writes OSC_52_READ where breaking on capitals alone
gives OSC52_READ and reports a spurious "not in the typedef".

GhosttyPlatform.MacOS and .IOS are Macos and Ios now, matching
GHOSTTY_PLATFORM_MACOS and GHOSTTY_PLATFORM_IOS. Neither is referenced on
Windows; only Windows is.
