+++ windows/Ghostty/Interop/NativeMethods.cs
@@ -19,85 +19,6 @@ internal readonly record struct GhosttyConfig(IntPtr Handle);
 internal readonly record struct GhosttySurface(IntPtr Handle);
 internal readonly record struct GhosttyInspector(IntPtr Handle);
 
-internal enum GhosttyPlatform
-{
-    Invalid = 0,
-    MacOS = 1,
-    IOS = 2,
-    Windows = 3,
-}
-
-internal enum GhosttySurfaceContext
-{
-    Window = 0,
-    Tab = 1,
-    Split = 2,
-}
-
-internal enum GhosttyClipboard
-{
-    Standard = 0,
-    Selection = 1,
-}
-
-internal enum GhosttyClipboardRequest
-{
-    Paste = 0,
-    Osc52Read = 1,
-    Osc52Write = 2,
-}
-
-internal enum GhosttyMouseState
-{
-    Release = 0,
-    Press = 1,
-}
-
-internal enum GhosttyMouseButton
-{
-    Unknown = 0,
-    Left = 1,
-    Right = 2,
-    Middle = 3,
-    Four = 4,
-    Five = 5,
-    Six = 6,
-    Seven = 7,
-    Eight = 8,
-    Nine = 9,
-    Ten = 10,
-    Eleven = 11,
-}
-
-internal enum GhosttyColorScheme
-{
-    Light = 0,
-    Dark = 1,
-}
-
-[Flags]
-internal enum GhosttyMods
-{
-    None = 0,
-    Shift = 1 << 0,
-    Ctrl = 1 << 1,
-    Alt = 1 << 2,
-    Super = 1 << 3,
-    Caps = 1 << 4,
-    Num = 1 << 5,
-    ShiftRight = 1 << 6,
-    CtrlRight = 1 << 7,
-    AltRight = 1 << 8,
-    SuperRight = 1 << 9,
-}
-
-internal enum GhosttyInputAction
-{
-    Release = 0,
-    Press = 1,
-    Repeat = 2,
-}
-
 [StructLayout(LayoutKind.Sequential)]
 internal struct GhosttyInputKey
 {
@@ -196,23 +117,6 @@ internal struct GhosttyCells
     public ushort CursorRow;
     public ushort CursorCol;
     public byte CursorInViewport;
-}
-
-// Mirrors ghostty_point_tag_e.
-internal enum GhosttyPointTag
-{
-    Active = 0,
-    Viewport = 1,
-    Screen = 2,
-    Surface = 3,
-}
-
-// Mirrors ghostty_point_coord_e.
-internal enum GhosttyPointCoord
-{
-    Exact = 0,
-    TopLeft = 1,
-    BottomRight = 2,
 }
 
 // Mirrors ghostty_point_s.
